socgam21_group18-backend/socgam21_group18-database/flaskapp/app/requirements.txt
apscheduler==3.7.0
amply==0.1.4
aniso8601==9.0.1
bson==0.5.10
certifi==2021.5.30
chardet==3.0.4
click==8.0.1
Cython==0.29.23
decorator==4.4.2
dnspython==2.1.0
docutils==0.17.1
email-validator==1.1.3
Flask==2.0.1
flask-mongoengine==1.0.0
Flask-PyMongo==2.2.0
Flask-RESTful==0.3.9
Flask-WTF==0.15.1
ftfy==6.0
gunicorn==20.1.0
idna==2.7
importlib-metadata==4.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
metis==0.2a5
mongoengine==0.23.1
networkx==2.5.1
numpy==1.21.0
pandas==1.3.0
petname==2.6
PuLP==2.4
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
requests==2.20.1
scipy==1.7.0
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.24.3
wcwidth==0.2.5
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.4.1
